Implement On/Off actions for the all tables
